+++ spec/features/compacted_config_spec.rb
@@ -58,7 +58,7 @@ describe 'Compacted config' do
         validate 'db.creds.*', :string
       end
 
-      compacted_config = Qonfig::Compacted.new(init_from: data_set_based_config)
+      compacted_config = Qonfig::Compacted.build_from(data_set_based_config)
 
       # NOTE: check readers
       expect(compacted_config.test).to eq(true)
@@ -87,7 +87,7 @@ describe 'Compacted config' do
     end
 
     specify 'fails on incorrect datasets passed to constructor' do
-      expect { Qonfig::Compacted.new(init_from: Object.new) }.to raise_error(Qonfig::ArgumentError)
+      expect { Qonfig::Compacted.build_from(Object.new) }.to raise_error(Qonfig::ArgumentError)
     end
 
     specify 'inheritance works as expected' do